Move Patches out of Spree namespace
This names the patches such that they do not live in the Spree
namespace. It also fixes a Rubocop offense.
