+++ lib/patches/backend/controllers/solidus_auth_devise/admin/orders/customer_details_controller_patch.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+module SolidusAuthDevise
+  module Admin
+    module Orders
+      module CustomerDetailsControllerPatch
+        def self.prepended(base)
+          base.before_action :check_authorization
+        end
+
+        private
+
+        def check_authorization
+          load_order
+          session[:access_token] ||= params[:token]
+
+          resource = @order
+          action = params[:action].to_sym
+          action = :edit if action == :show # show route renders :edit for this controller
+
+          authorize! action, resource, session[:access_token]
+        end
+
+        Spree::Admin::Orders::CustomerDetailsController.prepend self
+      end
+    end
+  end
+end
