--- lib/patches/backend/controllers/spree/admin/orders/customer_details_controller_patch.rb
@@ -1,22 +1,0 @@
-# frozen_string_literal: true
-
-module Spree::Admin::Orders::CustomerDetailsControllerPatch
-  def self.prepended(base)
-    base.before_action :check_authorization
-  end
-
-  private
-
-  def check_authorization
-    load_order
-    session[:access_token] ||= params[:token]
-
-    resource = @order
-    action = params[:action].to_sym
-    action = :edit if action == :show # show route renders :edit for this controller
-
-    authorize! action, resource, session[:access_token]
-  end
-
-  Spree::Admin::Orders::CustomerDetailsController.prepend self
-end
