+++ lib/patches/frontend/controllers/solidus_auth_devise/checkout_controller_patch.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module Spree
+module SolidusAuthDevise
   module CheckoutControllerPatch
     def self.prepended(base)
       base.before_action :check_registration, except: [:registration, :update_registration]
